Completed the solution for challenge 6
Minor changes made in how the guesses for the key size were obtained.
Namely, finding the Hamming distance of every consecutive block of
<key_size> bytes and finding the average, instead of just comparing
the first two blocks. Additionally the ordering in the Binary heap storing
the Hamming distance and key size was wrong, so fixed that.

Created a method to crack the key a byte at a time, decomposing
the problem into a series of single byte XOR ciphers.

Created functions to decode the base64 string into a binary
buffer which can be operated on easier

Finally all the functions were put together to crack the provided
multi-byte XOR-encoded, base64-encoded ciphertext.

Associated unit tests were written too.